review and refactor

+++ chapter20/projects/01/01.c
@@ -23,7 +23,7 @@ union punning_union {
  *
  * Parameters:
  *   - n: A pointer to the memory location to print.
- *   - nbytes: The number of bytes to print.
+ *   - nbytes: The number of bytes to print (up to 8).
  *
  * The function iterates over each bit in the memory location and displays its 
  * binary value. It starts with the most significant bit and shifts a mask to 
@@ -31,12 +31,12 @@ union punning_union {
  * The function prints '1' if the corresponding bit is set, and '0' otherwise.
  * After every 4 bits (nibble), a space is added for readability.
  */
-void print_binary(size_t *n, int nbytes);
+void print_binary(size_t *n, size_t nbytes);
 
 int main(void)
 {
-	struct IEEE_float f; 
-	union punning_union u;
+    struct IEEE_float f; 
+    union punning_union u;
 
     f.sign = 1;
     f.exponent = 128;
@@ -46,7 +46,7 @@ int main(void)
 
     printf("%f\n", u.value);
 
-	print_binary((size_t *)&u, sizeof(u));
+    print_binary((size_t *)&u, sizeof(u));
 
     printf("^^        ^____________________________  fraction\n"
            "||_______  exponent\n"
@@ -55,35 +55,30 @@ int main(void)
     return 0;
 }
 
-void print_binary(size_t *n, int nbytes)
+void print_binary(size_t *n, size_t nbytes)
 {
-    int i;
-    /* The mask aids to start with the most significant bit */
+    size_t i;
     unsigned long mask;  
+    /* The mask aids to start with the most significant bit */
+
     switch (nbytes)
     {
-        case 1: mask = 0x80;
-            break;
-        case 2: mask = 0x8000;
-            break;
-        case 3: mask = 0x800000;
-            break;
-        case 4: mask = 0x80000000;
-            break;
-        case 8: mask = 0x8000000000000000;
-            break;
-        default:
-            break;
+        case 1: mask = 0x80;               break;
+        case 2: mask = 0x8000;             break;
+        case 3: mask = 0x800000;           break;
+        case 4: mask = 0x80000000;         break;
+        case 8: mask = 0x8000000000000000; break;
+        default:                           break;
     }
 
     for (i = 0; i < nbytes * 8; i++) 
-	{
+    {
         putchar((*n & mask) ? '1' : '0');
         if (i % 4 == 3) 
         {
             putchar(' ');  /* Add space after each 4 bits (nibble) */
         }
-        mask >>= 1;  /* Shift the mask to the right for the next bit */
+        mask >>= 1;        /* Shift the mask to the right for the next bit */
     }
     putchar('\n');
